Fine-tune screen ready for Navigation compose
